friday

+++ tanks/tanks/degradconvert.h
@@ -1,3 +1,3 @@
 #pragma once
-float degtorad(int angle);
-int radtodeg(float angle);
+double degtorad(int angle);
+int radtodeg(double angle);
